perf: split the design page
