+++ Composer/packages/client/src/pages/design/VisualPanelHeader.tsx
@@ -32,7 +32,7 @@ type BreadcrumbItem = {
   onClick?: () => void;
 };
 
-type VisualEditorHeaderProps = {
+type VisualPanelHeaderProps = {
   projectId: string;
   visible: boolean;
   showCode: boolean;
@@ -142,7 +142,7 @@ const useBreadcrumbs = (projectId: string, pluginConfig?: PluginConfig) => {
   return breadcrumbArray;
 };
 
-const VisualEditorHeader: React.FC<VisualEditorHeaderProps> = (props) => {
+const VisualPanelHeader: React.FC<VisualPanelHeaderProps> = React.memo((props) => {
   const { visible, showCode, projectId, onShowCodeClick, pluginConfig } = props;
   const breadcrumbs = useBreadcrumbs(projectId, pluginConfig);
 
@@ -173,6 +173,6 @@ const VisualEditorHeader: React.FC<VisualEditorHeaderProps> = (props) => {
       </div>
     </div>
   ) : null;
-};
+});
 
-export default VisualEditorHeader;
+export default VisualPanelHeader;
